refactor: WebController를 App 컴포넌트와 RestaurantStore로 리팩토링
