+++ docs/REFACTOR.md
@@ -3,3 +3,8 @@
 - [x] 실행 오류...
 - [x] 테스트 코드 분리 및 명세 자세히
 - [x] 웹 컴포넌트 - data-xxx 속성 활용방법 바꾸기 -> 전체 데이터가 아닌 sort 조건, category 필터 조건으로
+
+- [ ] data-attribute 개선
+- [x] controller 리팩토링
+  - [x] RestaurantStore.ts - 레스토랑 데이터 저장, localStorage
+  - [x] App.ts - 컴포넌트 렌더링, 이벤트
