+++ src/index.js
@@ -1,5 +1,5 @@
+import App from './App';
 import '../style.css';
-import WebController from './Controller/WebController';
 
 import Dropdown from './components/Dropdown';
 import RestaurantList from './components/RestaurantList';
@@ -9,4 +9,4 @@ window.customElements.define('drop-down', Dropdown, { extends: 'select' });
 window.customElements.define('restaurant-list', RestaurantList, { extends: 'ul' });
 window.customElements.define('restaurant-item', RestaurantCard, { extends: 'li' });
 
-new WebController().run();
+new App().run();
